Prepare mask for inclusion of multiple vendors
* Use shifts for mask and make it into ssize_t (so it gets arithmetic
  shift instead of logical)
* Remove hardcoded names of nvidia in variable names outside of nvidia
  file.

+++ include/nvtop/extract_gpuinfo_common.h
@@ -148,7 +148,7 @@ struct gpu_vendor {
   const char *(*last_error_string)(void);
 
   bool (*get_device_handles)(struct list_head *devices, unsigned *count,
-                             uint64_t *mask);
+                             ssize_t *mask);
 
   void (*populate_static_info)(struct gpu_info *gpu_info);
   void (*refresh_dynamic_info)(struct gpu_info *gpu_info);
